backend/pyproject.toml
[project]
name = "lovable-crm"
version = "0.1.0"
description = "CRM system for Lovable website redesign previews"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "python-multipart>=0.0.6",
    "apscheduler>=3.10.0",
    "httpx>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "jinja2>=3.1.0",
    "stripe>=5.0.0",
]
